feat(LiveView): Highlighting for properties filters
Specifically for properties set in the 'Has property' and 'Has property of type' filters.

This commit also fixes properties in inherited classes from being included.
We only ever care about properties from the SuperStruct linked list, not the ClassPrivate linked list.
This is a bug in the Unreal module that hasn't yet been fixed, so we have to compensate here.